Move WooPage to a separate class
